WaylandBackend: copying to clipboard
This forwards the internal clipboard to the host, but does not handle
the primary selection nor accessing the host's clipboard internally.

The source is recreated for every new copy, rather than being reused
when it is still valid. This doesn't seem to be necessary, but doing it
differently leads to clipboard managers not becoming aware of the reused
copies (under Sway, `wl-paste --watch` doesn't notice them), which at
least suggests that not reusing is more common.